Moved core classes to new package.  Added denied actions support to Decorators
