+++ graylog2-server/src/main/java/org/graylog2/system/entities/SystemEntityAction.java
@@ -14,8 +14,15 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog.events.processor;
+package org.graylog2.system.entities;
 
+/**
+ * An action that can be performed on a {@link SystemEntity}.  An action can be allowed or denied irrespective of user roles or permissions.
+ *
+ * <p>
+ * For example, a user may have the permission <b>anomaly_configuration:edit</b> which allows her/him to edit Anomaly Detectors.  However, we may not want users to edit Anomaly Detectors installed via content packs as these could be updated or removed in future versions.  The user could clone such an entity, which has no action restrictions, and is not controlled/manage by Graylog and thus has no danger of being modified in anyway by a future version.
+ * </p>
+ */
 public enum SystemEntityAction {
     EDIT,
     DELETE,
